Added camera system and fixed transforms
Cameras no longer break rendering, but there are some scaling issues now (things must be made huge to even be visible). I intend to use this rendering engine for pixel art related things, so I may scale it all down and try to then scale it up for the final result, but we'll see.